made torch format bfloat16

+++ src/run_llm_inference.py
@@ -62,7 +62,7 @@ def load_hf_model_and_tokenizer(
     if tokenizer.pad_token is None:
         tokenizer.pad_token = tokenizer.eos_token
 
-    model = AutoModelForCausalLM.from_pretrained(model_name)
+    model = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype=torch.bfloat16)
 
     if use_lora:
         if adapter_dir is None:
